+++ pytest.ini
@@ -10,6 +10,7 @@ norecursedirs =
 markers =
         actuator: Tests for actuator agent
         actuator_pubsub: Test for actuator agent.
+        actuator_unit: Unit tests for actuator agent
         agent: Testing for core agent operations.
         alert: Testing alerts from the health subsystem.
         auth: Testing for auth based code.
